LCD 시간 오류 수정
로그 너무 많은점 수정
모터 기본 반복주기 수정

+++ src/service/motor_interval.hpp
@@ -11,7 +11,7 @@
 class MotorInterval : public Service {
     private:
         //기간
-        uint16_t intervalTime;
+        uint32_t intervalTime;
 
         //켜지는 시간
         uint16_t intervalSpan;
